+++ src/main/java/com/nhat/project/service/PostService.java
@@ -1,22 +1,24 @@
 package com.nhat.project.service;
 
+import com.nhat.project.dto.comment.CommentDto;
+import com.nhat.project.dto.post.*;
 import com.nhat.project.entity.Comment;
 import com.nhat.project.entity.Post;
 import com.nhat.project.entity.UpvotePost;
 import com.nhat.project.entity.User;
-import com.nhat.project.entity.id.UpvotePostId;
 import com.nhat.project.exception.NotOwnerException;
-import com.nhat.project.exception.NotValidOperationException;
-import com.nhat.project.exception.PostNotFoundException;
+import com.nhat.project.exception.InvalidOperationException;
+import com.nhat.project.exception.not_found.PostNotFoundException;
 import com.nhat.project.repository.CommentRepository;
 import com.nhat.project.repository.PostRepository;
 import com.nhat.project.repository.UpvotePostRepository;
 import com.nhat.project.repository.UserRepository;
 import org.springframework.beans.factory.annotation.Autowired;
 import org.springframework.data.domain.PageRequest;
-import org.springframework.data.domain.Sort;
+import org.springframework.data.domain.Pageable;
 import org.springframework.stereotype.Service;
 
+import java.util.ArrayList;
 import java.util.List;
 
 @Service
@@ -38,34 +40,44 @@ public class PostService {
         post.setOwner(user);
         post.setUpvote(0);
         postRepository.save(post);
-        //TODO fix createDate
     }
 
-    public void update(int id, String content, User user) throws NotOwnerException{
+    public PostUpdateDto update(int id, PostCreateDto postUpdateDto, User user) throws NotOwnerException, PostNotFoundException{
         Post post = postRepository.findById(id);
+        if (post == null) throw new PostNotFoundException("Post not found");
         if (post.getOwner() == user){
-            post.setContent(content);
+            String oldTitle = post.getTitle();
+            String oldContent = post.getContent();
+            post.setTitle(postUpdateDto.getTitle());
+            post.setContent(postUpdateDto.getContent());
             postRepository.save(post);
+            return new PostUpdateDto(id,oldTitle,post.getTitle(),oldContent,post.getContent(),user.convertToDto());
         }
         else {
             throw new NotOwnerException("You can not edit thus post");
         }
-        //TODO fix updateDate
     }
-    public void delete(int id, User user) throws NotOwnerException{
+    public PostDeleteDto delete(int id, User user) throws NotOwnerException, PostNotFoundException{
         Post post = postRepository.findById(id);
+
+        if (post == null) throw new PostNotFoundException("Post not found");
+
         if (post.getOwner() == user){
             postRepository.delete(post);
+            return new PostDeleteDto(post.getTitle(), post.getContent(), user.convertToDto());
         }
         else{
             throw new NotOwnerException("You can not delete this post");
         }
     }
-    public void upvote(int id, User user, int vote) throws NotValidOperationException{
+    public PostUpvoteDto upvote(int id, User user, int vote) throws InvalidOperationException, PostNotFoundException{
         if (vote != UPVOTE && vote != DOWNVOTE) {
-            throw new NotValidOperationException("Vote must be UPVOTE or DOWNVOTE ");
+            throw new InvalidOperationException("Vote must be UPVOTE or DOWNVOTE ");
         }
-        UpvotePost upvotePost = upvotePostRepository.findByUpvotePostId(new UpvotePostId(user.getId(), id));
+
+        if (postRepository.findById(id) == null) throw new PostNotFoundException("Post not found");
+
+        UpvotePost upvotePost = upvotePostRepository.findByUpvotePostId(user.getId(), id);
         Post post = postRepository.findById(id);
         if (upvotePost == null) {
             post.setUpvote(post.getUpvote() + vote);
@@ -82,21 +94,41 @@ public class PostService {
                 upvotePostRepository.save(upvotePost);
             }
         }
+        return new PostUpvoteDto(id,vote,post.convertToDto());
     }
-    public List<Post> showPostByUpvote(int page){
-        return postRepository.findAllByUpvote(PageRequest.of(page - 1, 5));
-    }
-    public List<Post> showPostByDate(int page){
-        return postRepository.findAllByDate(PageRequest.of(page - 1, 5));
-    }
-    public Post findById(int id) throws PostNotFoundException {
+    public PostDto getPost(int id, int page, String sortType) throws PostNotFoundException, InvalidOperationException {
         Post post = postRepository.findById(id);
-        if (post == null) throw new PostNotFoundException("Post not found");
-        return postRepository.findById(id);
-    }
 
-    public List<Comment> showComment(int id){
-        List<Comment> comments = commentRepository.findByPost(id);
-        return comments;
+        if (post == null) throw new PostNotFoundException("Post not found");
+
+        if (sortType != "latest" && sortType !="upvote") throw new InvalidOperationException("Invalid sort type");
+
+        Pageable pageable = PageRequest.of(page,15);
+        List<Comment> comments;
+        if (sortType == "latest") comments = commentRepository.findByPostSortByDate(id,pageable);
+        else comments = commentRepository.findByPostSortByUpvote(id,pageable);
+        List<CommentDto> commentDtos = new ArrayList<>();
+        for (Comment comment:comments){
+            commentDtos.add(comment.convertToDto());
+        }
+
+        return new PostDto(
+                post.getId(),
+                post.getTitle(),
+                post.getContent(),
+                post.getOwner().convertToDto(),
+                post.getUpvote(),
+                post.getCreateDate(),
+                post.getUpdateDate(),
+                commentDtos);
+    }
+    public List<PostShowDto> showPost(int page){
+        Pageable pageable = PageRequest.of(page,5);
+        List<PostShowDto> postShowDtos = new ArrayList<PostShowDto>();
+        List<Post> posts = postRepository.findAllByDate(pageable);
+        for (Post post:posts){
+            postShowDtos.add(new PostShowDto(post.getTitle(),post.getContent(),post.getOwner().convertToDto(),post.getCreateDate(),post.getUpdateDate()));
+        }
+        return postShowDtos;
     }
 }
